recator:extract login layout to layouts/login.py
