+++ app32.py
@@ -12,6 +12,8 @@ import os
 import uuid
 import base64
 
+from layouts.login import build_login_layout
+
 from layouts.reports import build_date_report_layout
 
 
@@ -302,27 +304,6 @@ def build_remarks_display_component(case_id: int):
             style={"backgroundColor": DARK_THEME['colors']['dark'][6]}
         ) for _, row in remarks_df.iterrows()
     ]
-
-
-def build_login_layout():
-    return dbc.Container([
-        dbc.Row(dbc.Col(html.Div(id='login-alert'), md=6), justify="center"),
-        dbc.Row(dbc.Col(dbc.Card([
-            dbc.CardHeader(html.H4("Login", className="text-center")),
-            dbc.CardBody([
-                dmc.Text("Username", size="sm"), dmc.TextInput(id='login-username', required=True),
-                dmc.Text("Password", size="sm", mt="sm"), dmc.PasswordInput(id='login-password', required=True),
-                dmc.Button("Login", id='login-button', fullWidth=True, mt="xl"),
-                dmc.Button("Forgot Password?", id="forgot-password-link", variant="subtle", size="xs", mt="xs",
-                           fullWidth=True)
-            ])
-        ]), md=5), justify="center", className="mt-5"),
-        dbc.Modal([dbc.ModalHeader(dbc.ModalTitle("Password Reset")),
-                   dbc.ModalBody(
-                       "To reset your password, please contact an administrator who can set a temporary password for you."),
-                   dbc.ModalFooter(dmc.Button("Close", id="close-forgot-password-modal"))],
-                  id="forgot-password-modal", is_open=False, centered=True)
-    ], fluid=True)
 
 
 def build_user_management_layout(privilege: str):
